- Changed directory structure - Http Connector to more FP implementation - Removed gitignore for build version
