+++ .gitignore
@@ -1,4 +1,3 @@
 .idea
 node_modules
 playgrounds
-/lib/
