+++ lib/Transformers/JsonApi/JsonApi.js
@@ -1,0 +1,1 @@
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=void 0;var _paramsSerializer=_interopRequireDefault(require("./Request/paramsSerializer")),_responseNormalizer=_interopRequireDefault(require("./Response/responseNormalizer")),_requestSerializer=_interopRequireDefault(require("./Request/requestSerializer")),_urlSerializer=_interopRequireDefault(require("./Request/urlSerializer"));function _interopRequireDefault(a){return a&&a.__esModule?a:{default:a}}var _default={paramsSerializer:_paramsSerializer.default,responseNormalizer:_responseNormalizer.default,requestSerializer:_requestSerializer.default,urlSerializer:_urlSerializer.default};exports.default=_default,module.exports=exports.default;
