+++ lib/Transformers/JsonApi/Request/Params/serializeParams.js
@@ -1,0 +1,3 @@
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=serializeParams;/**
+ * @param {ParameterBag} parameterBag
+ */function serializeParams(a){return a.params||{}}module.exports=exports.default;
