+++ lib/Transformers/JsonApi/Request/paramsSerializer.js
@@ -1,0 +1,3 @@
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=paramsSerializer;var _stringify=_interopRequireDefault(require("qs/lib/stringify")),_serializeFilters=_interopRequireDefault(require("./Params/serializeFilters")),_serializeParams=_interopRequireDefault(require("./Params/serializeParams")),_serializeSorting=_interopRequireDefault(require("./Params/serializeSorting")),_serializePaging=_interopRequireDefault(require("./Params/serializePaging"));function _interopRequireDefault(a){return a&&a.__esModule?a:{default:a}}function _objectSpread(a){for(var b=1;b<arguments.length;b++){var c=null==arguments[b]?{}:arguments[b],d=Object.keys(c);"function"==typeof Object.getOwnPropertySymbols&&(d=d.concat(Object.getOwnPropertySymbols(c).filter(function(a){return Object.getOwnPropertyDescriptor(c,a).enumerable}))),d.forEach(function(b){_defineProperty(a,b,c[b])})}return a}function _defineProperty(a,b,c){return b in a?Object.defineProperty(a,b,{value:c,enumerable:!0,configurable:!0,writable:!0}):a[b]=c,a}/**
+ * @param {ParameterBag} parameterBag
+ */function paramsSerializer(a){return(0,_stringify.default)(_objectSpread({},(0,_serializeFilters.default)(a),(0,_serializePaging.default)(a),(0,_serializeSorting.default)(a),(0,_serializeParams.default)(a)))}module.exports=exports.default;
