+++ lib/Transformers/JsonApi/Request/requestSerializer.js
@@ -1,0 +1,4 @@
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=requestSerializer;/**
+ * @param {Object} data
+ * @param {Object} headers
+ */ /* eslint-disable-next-line no-unused-vars */function requestSerializer(a){return a}module.exports=exports.default;
