+++ lib/Transformers/JsonApi/Response/Resource/Relationship/relationshipApplyToData.js
@@ -1,0 +1,5 @@
+"use strict";var _relationshipGetResource=_interopRequireDefault(require("./relationshipGetResource"));Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=relationshipApplyToData;function _interopRequireDefault(a){return a&&a.__esModule?a:{default:a}}function _slicedToArray(a,b){return _arrayWithHoles(a)||_iterableToArrayLimit(a,b)||_nonIterableRest()}function _nonIterableRest(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}function _iterableToArrayLimit(a,b){var c=[],d=!0,e=!1,f=void 0;try{for(var g,h=a[Symbol.iterator]();!(d=(g=h.next()).done)&&(c.push(g.value),!(b&&c.length===b));d=!0);}catch(a){e=!0,f=a}finally{try{d||null==h["return"]||h["return"]()}finally{if(e)throw f}}return c}function _arrayWithHoles(a){if(Array.isArray(a))return a}/**
+ * @param {Object<Resource[]>} items
+ * @param {Object<Resource[]>} includedRelations
+ */function relationshipApplyToData(a,b){Object.keys(a).forEach(function(c){Object.entries(a[c].relationships).forEach(function(d){var e=_slicedToArray(d,2),f=e[0],g=e[1];return null===g.data?void 0:g.isMany?void(// eslint-disable-next-line no-param-reassign
+a[c].data[f]=g.data.map(function(a){return(0,_relationshipGetResource.default)(a,b)})):void(a[c].data[f]=(0,_relationshipGetResource.default)(g.data,b))})})}module.exports=exports.default;
