+++ lib/index.js
@@ -1,0 +1,1 @@
+"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),Object.defineProperty(exports,"HttpConnector",{enumerable:!0,get:function get(){return _HttpConnector.default}}),Object.defineProperty(exports,"RepositoryManager",{enumerable:!0,get:function get(){return _RepositoryManager.default}}),Object.defineProperty(exports,"jsonApiTransformers",{enumerable:!0,get:function get(){return _JsonApi.default}}),Object.defineProperty(exports,"ParameterBag",{enumerable:!0,get:function get(){return _ParameterBag.default}}),Object.defineProperty(exports,"Collection",{enumerable:!0,get:function get(){return _Collection.default}});var _HttpConnector=_interopRequireDefault(require("./Connector/HttpConnector")),_RepositoryManager=_interopRequireDefault(require("./Repository/RepositoryManager")),_JsonApi=_interopRequireDefault(require("./Transformers/JsonApi/JsonApi")),_ParameterBag=_interopRequireDefault(require("./Resource/ParameterBag")),_Collection=_interopRequireDefault(require("./Resource/Collection"));function _interopRequireDefault(a){return a&&a.__esModule?a:{default:a}}
