+++ src/Transformers/JsonApi/Request/paramsSerializer.js
@@ -1,4 +1,4 @@
-import qs from 'qs';
+import stringify from 'qs/lib/stringify';
 import serializeFilters from './Params/serializeFilters';
 import serializeParams from './Params/serializeParams';
 import serializeSorting from './Params/serializeSorting';
@@ -8,7 +8,7 @@ import serializePaging from './Params/serializePaging';
  * @param {ParameterBag} parameterBag
  */
 export default function paramsSerializer(parameterBag) {
-  return qs.stringify({
+  return stringify({
     ...serializeFilters(parameterBag),
     ...serializePaging(parameterBag),
     ...serializeSorting(parameterBag),
