+++ src/Transformers/JsonApi/Response/Resource/Relationship/relationshipGetType.js
