+++ src/Transformers/JsonApi/Response/Resource/normalizeResources.js
