+++ test/unit/Transformers/JsonApi/paramsSerializer.test.js
@@ -1,9 +1,9 @@
 import ParameterBag from '../../../../src/Resource/ParameterBag';
-import paramsSerializer from '../../../../src/Transformers/JsonApi/paramsSerializer';
-import serializePaging from '../../../../src/Transformers/JsonApi/Params/serializePaging';
-import serializeFilters from '../../../../src/Transformers/JsonApi/Params/serializeFilters';
-import serializeParams from '../../../../src/Transformers/JsonApi/Params/serializeParams';
-import serializeSorting from '../../../../src/Transformers/JsonApi/Params/serializeSorting';
+import paramsSerializer from '../../../../src/Transformers/JsonApi/Request/paramsSerializer';
+import serializePaging from '../../../../src/Transformers/JsonApi/Request/Params/serializePaging';
+import serializeFilters from '../../../../src/Transformers/JsonApi/Request/Params/serializeFilters';
+import serializeParams from '../../../../src/Transformers/JsonApi/Request/Params/serializeParams';
+import serializeSorting from '../../../../src/Transformers/JsonApi/Request/Params/serializeSorting';
 
 describe('paramSerializer', () => {
   test('that serializeParams gives a valid JsonApi response', () => {
